Refactor CLI architecture to modularize cli components
