+++ src/cli/settings/drop.rs
@@ -1,0 +1,19 @@
+use clap::Parser;
+use serde::{Deserialize, Serialize};
+use crate::network::types::Probability;
+
+#[derive(Parser, Debug, Serialize, Deserialize)]
+pub struct DropOptions {
+    /// Probability of dropping packets, ranging from 0.0 to 1.0
+    #[arg(long = "drop-probability", id = "drop-probability")]
+    #[serde(default)]
+    pub probability: Option<Probability>,
+}
+
+impl Default for DropOptions {
+    fn default() -> Self {
+        DropOptions {
+            probability: Some(Probability::new(0.0).unwrap()),
+        }
+    }
+}
