+++ src/cli/settings/mod.rs
@@ -1,0 +1,8 @@
+pub mod packet_manipulation;
+mod drop;
+mod delay;
+mod throttle;
+mod reorder;
+mod tamper;
+mod duplicate;
+mod bandwidth;
