+++ src/cli/utils/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod serialization;
+pub mod logging;
